feat: implement sum of embedded cluster artifacts
makes a function to count all embedded cluster present in an airgap
definition. this function will be used later on when giving user feedback
about the amount of artifacts we are pushing.

as we are pushing different kinds of artifacts in different parts of the
kots source code it makes sense to implement this as part of the
embedded cluster artifacts struct.